+++ docs/changelog.rst
@@ -21,6 +21,13 @@ Master
     - Bug fixes
         - Fix websocket reconnection event.
 
+- ext.commands
+    - Additions
+        - Added support for the following typing constructs in command signatures:
+            - ``Union[A, B]`` / ``A | B``
+            - ``Optional[T]`` / ``T | None``
+            - ``Annotated[T, converter]`` (accessible through the ``typing_extensions`` module on older python versions)
+
 
 2.7.0
 ======