+++ docs/changelog.rst
@@ -3,8 +3,30 @@
 Master
 ======
 - TwitchIO
+    - Additions
+        - Added the new follower / followed endpoints for :class:`~twitchio.PartialUser`:
+            - :func:`~twitchio.PartialUser.fetch_channel_followers`
+            - :func:`~twitchio.PartialUser.fetch_channel_following`
+            - :func:`~twitchio.PartialUser.fetch_channel_follower_count`
+            - :func:`~twitchio.PartialUser.fetch_channel_following_count`
+        - The deprecated methods have had warnings added in the docs.
+        - New models for the new methods have been added:
+            - :class:`~twitchio.ChannelFollowerEvent`
+            - :class:`~twitchio.ChannelFollowingEvent`
+
     - Bug fixes
         - Fix IndexError when getting prefix when empty message is sent in a reply.
+
+- ext.eventsub
+    - Bug fixes
+        - Fix websocket reconnection event.
+
+- ext.commands
+    - Additions
+        - Added support for the following typing constructs in command signatures:
+            - ``Union[A, B]`` / ``A | B``
+            - ``Optional[T]`` / ``T | None``
+            - ``Annotated[T, converter]`` (accessible through the ``typing_extensions`` module on older python versions)
 
 - Docs
     - Added walkthrough for ext.commands