+++ docs/changelog.rst
@@ -13,6 +13,8 @@ Master
         - New models for the new methods have been added:
             - :class:`~twitchio.ChannelFollowerEvent`
             - :class:`~twitchio.ChannelFollowingEvent`
+        - New optional ``is_featured`` query parameter for :func:`~twitchio.PartialUser.fetch_clips` 
+        - New attribute :attr:`~twitchio.Clip.is_featured` for :class:`~twitchio.Clip`
 
     - Bug fixes
         - Fix IndexError when getting prefix when empty message is sent in a reply.