+++ docs/changelog.rst
@@ -24,6 +24,13 @@ Master
         - Fix websocket reconnection event.
         - Fix another websocket reconnect issue where it tried to decode nonexistent headers.
 
+- ext.commands
+    - Additions
+        - Added support for the following typing constructs in command signatures:
+            - ``Union[A, B]`` / ``A | B``
+            - ``Optional[T]`` / ``T | None``
+            - ``Annotated[T, converter]`` (accessible through the ``typing_extensions`` module on older python versions)
+
 
 2.7.0
 ======