+++ docs/changelog.rst
@@ -13,6 +13,8 @@ Master
         - New models for the new methods have been added:
             - :class:`~twitchio.ChannelFollowerEvent`
             - :class:`~twitchio.ChannelFollowingEvent`
+        - New optional ``is_featured`` query parameter for :func:`~twitchio.PartialUser.fetch_clips` 
+        - New attribute :attr:`~twitchio.Clip.is_featured` for :class:`~twitchio.Clip`
 
     - Bug fixes
         - Fix IndexError when getting prefix when empty message is sent in a reply.
@@ -20,6 +22,14 @@ Master
 - ext.eventsub
     - Bug fixes
         - Fix websocket reconnection event.
+        - Fix another websocket reconnect issue where it tried to decode nonexistent headers.
+
+- ext.commands
+    - Additions
+        - Added support for the following typing constructs in command signatures:
+            - ``Union[A, B]`` / ``A | B``
+            - ``Optional[T]`` / ``T | None``
+            - ``Annotated[T, converter]`` (accessible through the ``typing_extensions`` module on older python versions)
 
 - ext.commands
     - Additions